refactor: splits http server setup and wrapper
